Start using spire where we can.
One of the benefits of not being scala is that I don't have
to NIH everything. Spire is one of the finer scala libraries
available and instead of continuing to reinvent it I would do
better to exploit it. It's a very large dependency at the
moment, but one way or another we'll shrink it to the parts we
actually depend upon. For now we'll accept it in its entirety.

Our first contestants:

  type BooleanAlgebra[R] = spire.algebra.Bool[R]
  type UInt              = spire.math.UInt
  type Natural           = spire.math.Natural