Revert "For no fathomable reason, after we change to using ConsecutiveInt"
This reverts commit 2b40a7b8d2768ad831168982f6dfc9fd6b247a11.

Thanks to @dwijnand for figuring out that we had fallen into
an infinite loop on foreachReverse, because it was calling the
foreachReverse method on IndexRange, but I removed that, but
the implicit conversion took us right back into the caller.

That's a perfect example of the kind of disaster that is
modern programming: there are so many kinds of changes which
create a void into which some unexpected thing may step.
It's obviously not what the human intended, but the compiler
blithely proceeds, like there's no way it could know this
implicit conversion from out of left field which creates an
unbreakable loop wasn't precisely what I intended.