Drop some implicits.
Some level of implicitry will have to return but it's easier to
assess the total situation from a position of high explicitness.
It would be wonderful if scala were ever to offer us a way to
write implicits which act on literals only: I'd like the literal
number 8 to be usable as a Size without indiscriminate promotion
of every Int to Size. Maybe macros, or SIP-23 someday.