Overhaul all things pair and zip.
Major elements of this (far too large, but I need for you to
deal) patch.

1) Try to use -> pervasively in preference to tuple syntax
It's a type alias for Product2. This is important so everything
isn't needlessly overconstrained to being tuples. For instance
one should be able to zip anything which implements Product2.
Scala makes you box every element into an actual allocation of
Tuple2 before you can do anything with the zipped values.

2) Make ZipView more of a first class notion. A View[A1 -> A2]
can be transformed into a ZipView[A1, A2] but the one isn't the
other. All the operations on ZipView are written with the
knowledge that each element is a pair. This allows for a far
more convenient interface rather than the perpetual wrapping
and unwrapping which comes with scala's. For instance, which
of scala's two ugly babies would you like to adopt:

```
/* scala1 */  Seq(1 -> 2) map { case (x, y) => (x * 10, y) }
/* scala2 */  Seq(1 -> 2) map (x => (x._1 * 10, x._2))
/* psp    */  Seq(1 -> 2).zipView mapLeft (_ * 10)
```

3) Various renamings. Predicate is now ToBool, which has a
more obvious meaning, similarly ToString, ToInt, ToUnit,
and ToSelf replace less consistent and more jargonized names.
Names aren't done fluctuating yet either.