Hurtling toward release.
Another overlarge commit. Higher standards on that point
will come someday, for now we will bite the bullet.

- Separated Each from View, i.e. View no longer extends Each.
This is a step toward a near future when there is a strong
distinction maintained between non-views and views. Some
temporary scaffolding assists in the break-up (toEach, a
couple implicits, a sprinkling of .m calls) which will be
eliminated as the outlines of the "MaybeView" ADT take shape.

- Eliminated PartialOrder. Kick it downfield. We can use
Spire's if we need one.

- Worked over the API a few times. Its virtue is higher.

- Turned PairUp/PairDown into something more likely to be
understood, Pair.Split and Pair.Join.

- Simplified Size. Made newArray take an Int.