Imported scala library vector.
One of those rare bits of code in the scala library which
is genuinely useful. Amazingly it pretty much worked out of
the box after tearing out its entire inheritance hierarchy
and replacing it with... nothing.

I specialized it on a lark. Since that too appears to have
been successful I may now specialize the api interfaces.

The only remaining scala library code which I have any
interest in are the immutable HashSet/HashMap implementations,
so those may follow.