some returns an Option.
There's a tedious issue with type inference on Some(x)
expressions. On the one hand you are often in situations
where you want the type inferred as Option, such as when
there are multiple parameter lists and the type of the
first argument will constrain the type of the second one.

On the other hand, the return type of an extractor must be
Some[A] and not Option[A] if you wish to inform the compiler
that the extractor will always match, thus allowing the
possibility of exhaustivity checking.

Now I have some returning Option and Some returning Some.