Eliminating package object.
Package objects are misery. This is about the 100th bug
I've reported against package objects. Hopefully the last.

  https://issues.scala-lang.org/browse/SI-9580

That bug destroys any hope of separating interface and
implementation when package objects are involved, so let's
just forget the whole thing.

As always, this change reveals a bunch of inexplicable
scala bugs, especially with type inference.

[error] could not find implicit value for parameter z: UnbuildsAs[this.Elem, jMap[String,Int]]
[error]         jmap.m build,
[error]         ^

The "this.Elem" is a typical scala substitution bug.